# api-specs/openapi-spec-20251021-190000.yaml
openapi: 3.1.0
info:
  title: Decentralized Autonomous Forum API
  version: 1.0.0
  description: |
    RESTful API for the Decentralized Autonomous Forum platform. Features:
    - Point economy with crypto rewards
    - User authentication (email + OAuth2)
    - Post/comment creation with IPFS storage
    - Reputation-based moderation
    - BNB Chain integration

    **Base URL**: `https://api.forum.example.com/v1`

    **Authentication**: Session-based (HttpOnly cookies) + JWT for API clients

    **Rate Limiting**:
    - Anonymous: 100 req/hour
    - Authenticated: 1000 req/hour
    - Authenticated (Level 3+): 5000 req/hour

  contact:
    name: API Support
    email: support@forum.example.com

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.forum.example.com/v1
    description: Production server
  - url: https://staging-api.forum.example.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie (HttpOnly, Secure, SameSite=Lax)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API clients

# ============================================================================
# REUSABLE SCHEMAS
# ============================================================================
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[A-Za-z0-9_-]+$'
          example: crypto_enthusiast
        display_name:
          type: string
          maxLength: 100
          example: Crypto Enthusiast
        bio:
          type: string
          maxLength: 500
          example: "Passionate about DeFi and BNB Chain"
        avatar_url:
          type: string
          format: uri
          example: "https://ipfs.io/ipfs/Qm..."
        points:
          type: integer
          minimum: 0
          example: 1250
        level:
          type: string
          enum: [new_user, active_user, trusted_user, moderator, senior_moderator, admin]
          example: active_user
        total_posts_created:
          type: integer
          example: 42
        total_comments_created:
          type: integer
          example: 128
        total_likes_received:
          type: integer
          example: 456
        is_verified:
          type: boolean
          example: false
        bnb_wallet_address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
        created_at:
          type: string
          format: date-time
          example: "2025-10-21T09:25:00Z"
      required:
        - id
        - username
        - points
        - level
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
      required:
        - error
        - message

    # ========== Pagination Schemas ==========
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user with email
      description: |
        Create a new user account with email and password.
        - Password must be at least 8 characters
        - Username must be 3-50 alphanumeric characters
        - Receives 100 bonus points upon email verification
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  pattern: '^[A-Za-z0-9_-]+$'
                password:
                  type: string
                  minLength: 8
                  format: password
              required:
                - email
                - username
                - password
            example:
              email: "newuser@example.com"
              username: "crypto_newbie"
              password: "SecurePass123!"
      responses:
        '201':
          description: User created successfully, verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Registration successful. Please check your email to verify your account."
                  user_id:
                    type: integer
                    example: 12345
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (max 5 registrations per hour per IP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email and password
      description: |
        Authenticate user and create session.
        - Returns HttpOnly session cookie
        - Session expires after 7 days of inactivity
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
            example:
              email: "user@example.com"
              password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "session_id=abc123; Path=/; HttpOnly; Secure; SameSite=Lax; Max-Age=604800"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (max 5 login attempts per 15 minutes per IP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout current user
      description: Invalidate session and clear session cookie
      operationId: logoutUser
      security:
        - SessionAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"
  /auth/oauth/{provider}:
    get:
      tags:
        - Authentication
      summary: OAuth2 login redirect
      description: |
        Redirect to OAuth2 provider for authentication.
        Supported providers: discord, twitter, telegram, reddit, facebook
      operationId: oauthLogin
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [discord, twitter, telegram, reddit, facebook]
      responses:
        '302':
          description: Redirect to OAuth2 provider
  /auth/oauth/{provider}/callback:
    get:
      tags:
        - Authentication
      summary: OAuth2 callback
      description: Handle OAuth2 provider callback and create session
      operationId: oauthCallback
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [discord, twitter, telegram, reddit, facebook]
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to home page with session cookie

  # ========== User Endpoints ==========
...